service refactor

+++ src/main/java/com/prueba/gestionriesgos/domain/dto/PersonDTO.java
@@ -19,7 +19,7 @@ public class PersonDTO {
      * The unique identifier for the person.
      * This serves as the primary key to uniquely identify a person in the database.
      */
-    private Integer personId;
+    private Long personId;
 
     /**
      * The first name of the person.
